requirements.txt
jupyterlab
numpy
pandas
matplotlib
seaborn
streamlit==1.26.0
python-dotenv==1.0.0
folium
streamlit_folium
geopandas